Tested working with Code Blocks Arduino edition version 1/31/2013
Fixed up throttle code, made it verify my accelerator pedal output. This would be a setup where there are two pedal pots, one is to be twice the value of the other. If your pedal doesn't match then adjust accordingly. Also, regen is disabled.

Added a mode to the menu where the accelerator can be used to drive the motor instead of merely doing an RPM ramp.