Renamed some things
Moved ENUMs into the classes they belong with to better scope them

Changed the parameters of functions to enums where appropriate to enforce proper values

Added status reporting for pot based throttle so that it reports what caused an error

Pretty much totally reworked how the throttle code works so that it properly handles throttles which go either way (high to low or low to high). Now pretty much any pot based throttle should be able to be used. Initial caution would, however, be advisable when testing a new pedal.

Updated the readme to explain how to get started with the project.