Switch to using SerialUSB instead of Serial
Changed all references to Serial to SerialUSB so that all debugging output
runs over the normal USB connector instead of the debug connector. Why?
Because connecting a PC to Serial reboots the Due. This is bad when you need
the Due to be processing canbus frames for the DMOC. So, this change
allows the Due to operate unimpeded. It also means you won't see the
initial debugging output unless you push reset yourself.