Behold! A Visual Studio 2012 project for GEVCU.
No, this doesn't mean you can compile the firmware in VS2012 but you can view it and edit it there. Here is what you want to do: Edit with VS2012 if that's your thing. Then, in the Arduino IDE File->Preferences->Use External Editor. Now you can edit with the external editor of choice and compile/upload with the Arduino IDE. It will all magically work. You're welcome for not having to use the craptastic Arduino IDE anymore.