Added Wrapper class and pre-compiler checks to be able to support Uno and Due based platforms.
Added CanHandler (can_handler.c) which wraps the Due CAN and the machina CAN and moved all implementation
specific code there (mainly from GEVCU.ino and DMOC.cpp).
* Added GEVCU.h to move definitions and includes out of the main code.
* Added config.h to define global definitions centralls (e.g. pin assignments)
* Added some code in GEVCU.ino to monitor values on an LCD screen (will be moved to a separate class)
* Re-Added Machina specific code to pref_handler.cpp and timer.cpp
* Added .project and .cproject files so it can be loaded into eclipse as an existing project with the plugin from http://www.baeyens.it/eclipse/V2 (much better than the stock IDE).

Note: The code compiles fine for both platforms in Eclipse (with http://www.baeyens.it/eclipse/V2)
and the Arduino IDE. But the binaries were not verified on the boards yet (as I still lack a CAN shield
for the Due and a Machina).