Proper two throttle pot support
Made some changes to the throttle calibration to fix issues since
refactoring. Also added output of percentage of fluctuation of pedal
values at rest and full throttle during calibration.
Changed the pot throttle to use a percentage threshold for validation
and not raw values
Implemented support for two throttles with the second inverted
Increased the config thresholds for invalid values since I am getting
in excess of 10% fluctuation in values