Added a super secret command X - this will essentially cause a software reset without disconnecting the USB serial.
The potential utility for this command while debugging over SerialUSB should be obvious. What should also be obvious is that you should never use this command while connected to any important external hardware. The command just calls setup() again which reinitializes everything. But, the old data is not deallocated first. This is potentially dangerous (hence don't do it with stuff connected) but allows for developers to see the debugging text on start up even when connected over SerialUSB.

It would likely be wise to remove this command in builds destined for general use.