Hard coded the brake ADC temporarily.
The long story: Throttle and brake ADC have been moved to EEPROM. But, the way the potbrake code reads the EEPROM leads to old defaults being used when you've had the board a while. These old defaults then set the brake to use the first ADC pin which is wrong. So it is hard coded so that people don't get a big surprise. New boards will never have this problem.