Fixed Operating Time and Web Page Variable Display
Web page would often not load until a change was made in serial port.
Operating time was not displaying.  Lots of little  fixes.
Changed torqueRequested in CodaMotorController to NOT include the
offset.  Need to do this in DMOC Motor Controller as well.  This is
because the offsets, which we were taking out in ICHIP are different
between the two, and likely from one inverter to the next.