Lots of MotorController
Promoted Operation State (enable/disable) and gears (forward/reverse)
to parent MotorController and eliminated them in childs DMOC645 and Coda

added code for a reverse input that will put the device in reverse

Added code for an enable input that can be used to put a device into
disabled state

added brakelight output to turn on brakelights during regen.

added a reverse light output to activate reverse lights when were in
reverse gear.