Revised Max RPM use
Max RPM on the config screen, as it turns out, was only used to set the
speed command in the DMOC in speed mode.

The web page implies we are setting a maximum RPM which will not be
exceeded yet there is nothing in eithe CodaMotorController or
DmocMotorController to implement this.

Added code to send a ZERO torque command in the event the inverter
reports an actual speed greater than the maximum rpm.

This should prevent running the motor above max rpm.