Running Lights
I’ve repurposed the RUNNING light on the STATUS panel and the running
variable to indicate that we are receiving frames from the
inverter/controller.  If no frames are received for a certain time, the
light goes out.  If we never receive frames, it never comes on.

In this way, we will know we are communicating or at least receiving
traffic if we have a RUNNING advisory on the STATUS page.